+++ packages/components/src/components.d.ts
@@ -185,6 +185,10 @@ export namespace Components {
     }
     interface MiDropdown {
         /**
+          * Clear filter.
+         */
+        "clearFilter": () => Promise<void>;
+        /**
           * Sets the disabled state for the dropdown.
          */
         "disabled": boolean;
@@ -231,7 +235,12 @@ export namespace Components {
          */
         "open": boolean;
         /**
-          * Gets the selected items
+          * If present, it dictates placeholder for an filterable input field in the dropdown. Defaults to 'Type to filter...'.
+          * @type {string}
+         */
+        "placeholder": string;
+        /**
+          * Gets the selected items.
           * @type {Array<HTMLMiDropdownItemElement>}
          */
         "selected": Array<HTMLMiDropdownItemElement>;
@@ -863,6 +872,10 @@ export namespace Components {
           * @type {number}
          */
         "scrollLength": number;
+        /**
+          * Update scroll buttons enabled/disabled states.
+         */
+        "updateScrollButtons": () => Promise<any>;
         /**
           * Updates enable/disable state for scroll up and down buttons.
           * @returns
@@ -1608,12 +1621,21 @@ declare namespace LocalJSX {
          */
         "onChange"?: (event: MiDropdownCustomEvent<any>) => void;
         /**
+          * Emit an event when search field is cleared.
+         */
+        "onCleared"?: (event: MiDropdownCustomEvent<void>) => void;
+        /**
           * Gets or sets the state of the dropdown. If the attribute is set to true then the dropdown will be expanded.
           * @type {boolean}
          */
         "open"?: boolean;
         /**
-          * Gets the selected items
+          * If present, it dictates placeholder for an filterable input field in the dropdown. Defaults to 'Type to filter...'.
+          * @type {string}
+         */
+        "placeholder"?: string;
+        /**
+          * Gets the selected items.
           * @type {Array<HTMLMiDropdownItemElement>}
          */
         "selected"?: Array<HTMLMiDropdownItemElement>;