+++ packages/map-template/src/atoms/venuesState.js
@@ -1,8 +1,0 @@
-import { atom } from 'recoil';
-
-const venuesState = atom({
-    key: 'venues',
-    default: []
-});
-
-export default venuesState;